[Findit] Modifying file path line regular expression pattern to account for additional cases.
Presently line numbers in the form file.extension(<line number>) and file.extension:<line number> are supported. This change adds support for the case file.extension:[line <line number>] as found in cases for FindBugs failures for Android builds. Without this change, findit is only able to extract the filename, but not the line number, which can lead to false positives should we add support for filtering by line distance in the future.

BUG=506360

Review URL: https://codereview.chromium.org/1277293002