[Findit] Implementing a Junit test signal extractor and re factoring InstrumentationTestExtractor
This change adds support for Junit test failures which are in a different format from InstrumentationTests. Since much of the logic to convert a java package name to a java file, some re-factoring work to share the code between the two extractors is also included in this change.

BUG=524842

Review URL: https://codereview.chromium.org/1327613003