Fix gif selfie feature.
MediaStream.stop() is now MediaStreamTrack.stop() so we need to getTracks()
and stop the first track (the video one). The worker script is also hosted
from the root, so we need to start with a slash otherwise the file is
loaded relative to the issue url (/123/).

Review URL: https://codereview.chromium.org/1382993003