Don't party while ct-sheriff-o-matic is waiting for a response.
Currently, when one navigates to https://sheriff-o-matic.appspot.com/chromium, there's a flash of party balloons while the failures are loading.

This CL would change it so that whenever ct-sheriff-o-matic is loading, no party balloons are shown in ct-unexpected-failures.

Is there a better way to do this? Are there any other related changes that could be made?

Review URL: https://codereview.chromium.org/1394273002