Run testjs.py when js files are modified
This CL causes the infra_repo_trybot recipe to actually run the
js tests (as specified by testjs.py) in the infra.git repo.
It doesn't run them on Windows, because.
I also changed the names of some of the test steps to be moar bettar.

R=hinoka@chromium.org
BUG=547987

Review URL: https://codereview.chromium.org/1414743009

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297328 0039d316-1c4b-4281-b951-d872f2087c98