[Findit] Adapting findit to use new wf_config when determining support level for masters
Now that the findit config page is built, the parts of the code that were checking for masters' support levels can now be adapted to use it. This change removes waterfall_config (which was previously being used only for supported masters) and replaces it with masters.py, which hooks into the new wf_config created as part of https://codereview.chromium.org/1409753005 that handles saving the settings to the datastore.

BUG=542932

Review URL: https://codereview.chromium.org/1447823002