[Findit] Temporary fix for config import break.
For some reason, the line

from components.datastore_utils import config

causes import errors in other parts of the code. This change removes the dependency on config.py found in infra/appengine/luci/, though with many shared pieces. Eventually this CL should be either revisited or reverted when a fix for the import error is implemented in order to reuse as much of luci's config.py as possible.

BUG=564356

Review URL: https://codereview.chromium.org/1490983002