Shard cumulative metrics across multiple memcache keys.
This abuses the TaskTarget's task_num field to hold the shard number.  The app doesn't attempt to sum the counts across shards - it just exports a separate stream for each shard to be aggregated later.  This ensures that stream resets (when memcache is partially flushed) are handled correctly.

BUG=541411

Review URL: https://codereview.chromium.org/1491383004