infra_libs/logs: Check if log directory exists.
Check if a log directory exists before probing it and attempting to
create a temporary file. This check prevents a condition being
encountered on Windows 10 where E:\ is mapped to a removable disk and,
consqeuently, the attempt to blindly create a file on it causes the
script to hang pending removable media insertion.

BUG=chromium:566094
TEST=local
  - Ran a test case on the afflicted system, hangs forever.
  - Applied fix, system works as expected.

Review URL: https://codereview.chromium.org/1527843002