Mark issue as updated upon creation
Previously we've not marked issue as updated on purpose, so that an update with
the links to the flakes is posted at the next run of the update_issue_tracker
cron job. Since we now do not post links anymore and also report number of
recent flakes in the message posted when creating issue, we should also mark the
issue as updated.

R=akuegel@chromium.org
BUG=556840

Review URL: https://codereview.chromium.org/1532623002