[Findit] Fixing target extraction on non-windows platform to account for spaces and quotes in target
In some cases such as https://findit-for-me.appspot.com/build-failure?url=https://build.chromium.org/p/chromium.mac/builders/Mac%20Builder/builds/33499, the failed target may have quotes and spaces in the name. If this is the case, the previous regex will match the first quote but then stop at the first space, giving an incomplete target. This change accounts for this case, and adds the full quoted name along with a comment and alternate regex if the quotes are undesired as part of the target to be passed.

BUG=576528

Review URL: https://codereview.chromium.org/1571333006