[sheriff-o-matic] Workaround for missing menu after updating to chrome 49.
I poked around in the developer tools to see if I could fix it in styling but
this looks to be an issue with core-scaffolding.

Visit this page with 49 and you won't see the left nav either. Even invoking
the responsive layout change (make the window thinner until the drawer
hides itself under the hamburger menu in the upper left) won't make the
menu items appear:

https://www.polymer-project.org/0.5/components/core-scaffold/demo.html

So I removed scaffolding and fixed the layout with css. It isn't as pretty
but it's visible at least. Thankfully I didn't have to giveupandusetables.com.

BUG=583118

Review URL: https://codereview.chromium.org/1657173002