[Issue wizard] Call encode() correctly for either unicode or str.
This is a fix to a bad fix.  In a CL back in November I added validation
for the continue parameter:
https://codereview.chromium.org/1473493002

But, there was a problem because the string was sometimes unicode and
could not be used in a redirect header.  I tried to fix it in CL:
https://codereview.chromium.org/1477063002

But, that failed for the non-unicode because of a coding error.  Here
is the right way to implement it.  I tested locally with some URLs that
were plain text and some that included unicode chars.

R=sheyang
BUG=534790

Review URL: https://codereview.chromium.org/1680563003