[Findit] Trigger a new Swarming task to re-run selected tests in a referred completed Swarming task.
The main purpose of this CL is to differentiate reliable test failures from flaky test failures on the main waterfall, by re-running the failed tests of a gtest step right after the failures are detected.

This is mainly for speed: we could avoid the time-consuming compile, because all the needed data (test binary + test data) are fresh in the isolated server or Swarming bots.

Main flow:
0. Retrieve a task id of any sharding Swarming task of the gtest step.
1. Retrieve the task request parameters.
2. Overwrite/update the parameters of the referred task request.
3. Trigger a new Swarming task with the new parameters.

BUG=583105

Review URL: https://codereview.chromium.org/1655023002