[Issue Wizard] Use components when posting to monorail.
This makes the wizard handle components appropriately for both
codesite and monorail.  If the issue is gocing to code.google.com,
then leave the labels as-is.  If it is going to any other site, assume
that is monorail, and if label1 is a Cr-* label then convert that to a
component.  None of the wizard options sets more than one Cr-* label
or does it with anything other than label1.

R=sheyang
BUG=775

Review URL: https://codereview.chromium.org/1680423002