Adding configurable properties for try job and swarming tasks to wf_config.
This change only adds the properties to the config, but does not make Findit depend on it. To prevent downtime, a new version of Findit with this change will have to be deployed first, its config set, then the code updated to rely on this config.

BUG=

Review URL: https://codereview.chromium.org/1713853002