[Rietveld] Convert newlines to br and urlize links in Rietveld HTML emails.
Context is in
https://groups.google.com/a/chromium.org/d/msg/infra-dev/KlvYUSACdSo/VWZtM5LxBgAJ
and
BUG=chromium:582463

I started off looking at making a user setting for receiving emails only in text but this turned out to be much messier than I expected because we concatenate the to and cc lists. We would have to separately send out emails for users who only wanted in text but then we would lost the CC list.

Instead I started looking into making the appropriate modifications to the HTML body because sending it out and this appears to work.

This code is staged here:
https://skia-codereview-staging.appspot.com/
Please feel free to play with it.
Eg: https://skia-codereview-staging.appspot.com/3071001/

Review URL: https://codereview.chromium.org/1692953002