Chromium-status: Resolve Git commit hashes to crrev.com URLs
Currently, https://chromium-status.appspot.com/ resolves Git commit hashes to
chromium-review.googlesource.com links that are broken most of the time, i.e.
in the primary use case when said commits are to chromium/src.

Change this so that the links point to crrev.com instead, which will redirect
to the respective commit in Gitiles' repository browser.

BUG=None

Review URL: https://codereview.chromium.org/1725923002