Fix cloudstorage in DEPS to use the real mirror
This one needs a real review, because it changes the package to use the full
client repo instead of a subtree mirror of it. I checked and the full repo is
only about 4MB, so I think the reduction in complexity is a worthwhile tradeoff.

R=agable@chromium.org
BUG=587260

Review URL: https://codereview.chromium.org/1773943002