Use UTC instead of local time almost everywhere in the cq_stats.py script
This doesn't actually change the stats as we are passing timestamp since epoch
to the chromium-cq-stats, which is timezone independent. This only changes how
time range is printed to the script and also makes reasoning about times used in
the scripts easier. There is still one call that deals with local time, which
converts time of the commit returned by Git from local to the UTC.

R=phajdan.jr@chromium.org, sergeyberezin@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1779513003