Only report bugs as stale if there were no updates for 7 days
This fixes the case when the bug was not updated for 30 days and was just moved
to Sheriff Bug Queue. Since we've considered third_party updates only for
computing 7 day deadline, we would have immediately reported it as stale, which
is not true, because Sheriffs haven't had a chance to look at it yet.

R=tandrii@chromium.org
BUG=545408

Review URL: https://codereview.chromium.org/1789543002