Set category and other non cTree dependent CtFailureGroup properties on construction.
Currently we set this.category and other properties after
CTTreeList.getTree completes. This seems unnecessary for most
properties and makes data binding in the UI more complicated.
this.data.bugLabels is the only property that depends on cTree.

Also fixes a small error when a tree list is empty.

BUG=

Review URL: https://codereview.chromium.org/1838363002