Switch to Glide for Go dependency management.
Goop is long abandoned. Glide on the other hand has a hip website and a logo
(http://glide.sh/), which clearly means it is superior to Goop in every way.
At least it has an active community.

We still had to introduce hacks to make it work in infra + infra_internal
environment. All the ugliness is hidden behind 'deps.py' script that acts as
a frontend for glide. In fact, glide is not exposed in any way and should be
considered an implementation detail at this point.

R=nodir, dnj
BUG=

Review URL: https://codereview.chromium.org/1844483002