[Monorail] Remove tiny ">" indicator on your issues and reduce padding.
After experimenting with the alignment, I took a step back to evaluate
the usefulness of the feature in the first place.  I don't think that
users will miss it.  It was most useful when the owner column is not
shown, but none of our projects configure columns that way.  If
users do miss it, let's find some new approach to visuallying
highlighting the isues that the current user owns.

R=agable@chromium.org
BUG=monorail:495

Review URL: https://codereview.chromium.org/1904263002