[Monorail] Limit autocomplete options for faster typing on mobile.
The autocomplete menu for text like "compone" was super-slow on
mobile because we were constructing a <table> with over 1200 rows.
The fix is to limit it to 100.  If the user does not find what they
are looking for in those 100, they will need to type more chars.
We are not limiting the choices considered, just the choices that
are added to the DOM on each keystroke.

With this change, my Nexus 5 is able to keep up with my typing
even when I am quickly presing backspace.

Preview:
https://3889-4167fab-tainted-jrobbins-dot-monorail-staging.appspot.com/p/chromium

R=agable@chromium.org
BUG=monorail:756

Review-Url: https://codereview.chromium.org/1921003007