[Findit] Use PubSub built-in authentication for push endpoint.
After b/25788912, subscription in the same project as the push endpoint could bypass the authentication as shown in this CL. The authentication approach is described in https://cloud.google.com/pubsub/prereqs.

This approach is verified through dummy apps on both appspot and googleplex.
That means Fracas could also do push delivery instead of pull.

BUG=600460

Review-Url: https://codereview.chromium.org/1940843003