[Monorail] Implement "Switch accounts" via google.com/a/SelectSession.
Changing the URL from '/accounts/ServiceLogin' to '/a/SelectSession'
always gives the user the chance to switch accounts, even if they
are currently signed into exactly one account.  Also, that account
selection page looks nicer.  The user will see both the google.com
and the GAE account selectors the first time for each account, but
it should not happen much after that if the user leaves "Remember
this choice for 30 days" checked.

R=seanmccullough@chromium.org
BUG=monorail:438

Review-Url: https://codereview.chromium.org/1954193002