Add monitoring for success/failure of connections to Cloud SQL
The Cloud SQL SLA is for 99.95% uptime, where uptime is defined as
"less than 20% of connection attempts failing". However, the Cloud
Console doesn't appear to actually monitor the connection error rate
for us. This will give us our own graph, so we can see if/when the
SLA is not being upheld.

R=jrobbins@google.com

Review-Url: https://codereview.chromium.org/1976973003