bot_start: Drop swarming signal file when running swarming
Swarming's bootstrap method assumes that it's solely responsible for startup, but
that's not true for chromebuild_startup.py based startup on CCompute.  This
drops a signal file so that swarming doesn't try to, otherwise they'll fight
for being the sole owner and weird things happen.

BUG=
TBR=maruel

Review URL: https://codereview.chromium.org/2038953002 .