[Monorail] Also pass global issue IDs to notification tasks.
This is the first of two CLs that will be needed to resolve the problem
of referencing issues that are moved when creating a notification task.
The parameters passed to notification tasks were in the form of project_id
and local issue IDs.  That is a poor choice left over from the days before
there were global issue IDs.  With this CL, we start passing both the global
and local issue IDs.  After this is deployed, the follow-up CL will acutally
make use of the global IDs and stop passing the local ones.

R=sheyang@chromium.org
BUG=monorail:1352

Review-Url: https://codereview.chromium.org/2039703002