[monorail] move feature extraction into a shared module so training and serving use the same feature sets.
I've run ROC analysis on this model, trained with the original codesite examples
as well as some of our new examples exported from prod over the last couple of months.
In staging and production the model is called '20160617'.  Here's the ROC curve/data:
https://docs.google.com/spreadsheets/d/1f5WkXyNDoVc22l4eTU1GBwnUBbrQDG7fVFVBX6HyKjU/edit#gid=939002793

It's a decent improvement over the current model.

Two new features added are raw length and gzipped length for the description and summary fields. The rationale is that many spammers repeat messages over and over so they should have high compression ratios compared to legit content.

Caveat:
This CL doesn't include "Unsafe URL count" as a feature because I don't see a
good way to access the Safe Browsing API key in production. We can't put it in
settings.py obviously, and GAE doesn't apparently have an API for accessing
keys like this.

BUG=monorail:1316

Review-Url: https://codereview.chromium.org/2076743003