[Monorail] Pass default user query strings explicitly to search backends in q= param.
The problem was that when the user hits a frontend with no q= parameter, that is interpreted as being a default query, e.g., [owner:me].  However, when the frontend hit the backends, it didn't pass any q=, which meant that the backends were supposed to follow the same logic.  However, backends do not handle requests in the context of a single project and therefore did not look up a config or the default query for any project, so no default query.   It is intentional that backend requests are not in the context of a single project so that we can handle cross-project searches for dashboards.

The fix is for the frontend to explicitly pass mr.query (which may be a default query) in the q= parameter to the search backends.

R=seanmccullough@chromium.org
BUG=monorail:1421

Review-Url: https://codereview.chromium.org/2082853002