[Monorail] Re-enable keyboard shortcuts on issue entry page.
The problem was that users who are fast typists could load the issue
entry page and start typing "u" and "?" and "/" before the browser
focus was set on the summary input field.  That happened on
code.google.com which set the focus using JS.

In monorail, the underlying problem was resolved by CL
https://codereview.chromium.org/1909363006
which set the autofocus attribute on the summary field.
Now that that is in place, the original keyboard shortcuts
code work safely (meaning that the user must click away from all
input fields before keystroke shortcuts are handled).

R=seanmccullough@chromium.org
BUG=monorail:481

Review-Url: https://codereview.chromium.org/2078333007