SoM: Limit number of test failures displayed per alert
Before, lots of tests listed for a single Sheriff-o-Matic alert could stretch the page vertically a lot, causing Sheriffs to have to scroll a lot to get to the next alert even if they are not interested in reading the test failures.

I tweaked the CSS a bit so that the test list for a single alert becomes a scrollable box if the test failures box exceeds a certain height.

BUG=

Review-Url: https://codereview.chromium.org/2095833002