[Monorail] Make issue notification tasks identify issues by issue_id.
This is the second of a series of three CLs that will change from using
(project_id, local_id) to just issue_id when specifying issues in tasks
that send email notifications.  I had previously thought I would do it
in 2 CLs, but I'll do it in 3 to make it safer in case we need to roll
back.  The upcoming 3rd CL will be the one that stops passing
project_id and local_id.

R=sheyang@chromium.org
BUG=monorail:1352

Review-Url: https://codereview.chromium.org/2099073003