[Findit] Fix flaky dashboard test.
The tests are flaky because of the request_time for testing fracas analyses are not fixed, utcnow() is used for testing, so if the tests are running right pass the midnight of today, the assertion would fail. Fix this by hard-coding the request time of testcases and setting start_date and end_date in the request.

BUG=624998, 626511

Review-Url: https://codereview.chromium.org/2147773002