test-results: Initial commit for Go service
Introduce a Go module to handle GET /testfile requests. The dispatch.yaml
defines rules to route requests to the /testfile path to the Go module.

Besides the deviations noted below, this commit aims
to be an exact port of the existing Python GET /testfile handler to Go,
to maintain compatibility with clients using the service.

* Return appropriate HTTP status codes.
* Remove the 'Delete' link next to each test file in the testfiles
  list page (estaab@).
* Do not output 'None' links for null build_numbers .
* The query param build_number=None is no longer ignored. Previously,
  this was needed because of the 'None' link described above.
* Remove optional tags in HTML

BUG=

Review-Url: https://codereview.chromium.org/2109363002