[Monorail] Update the URL used by "Switch accounts".
After some digging and experimenting, I found the new GAIA URL that works
exacty like the old one used to work.  I found it by visiting different google products until one of them presented me with a choice of accounts.

Even though this URL would work when the user is first signing in, I
am keeping the logic that uses the URL given to us by GAE for the
initial sign-in flow because I want to avoid any future breakage that
would prevent users from signing in to the first account.

R=agable@chromium.org
BUG=monorail:1563

Review-Url: https://codereview.chromium.org/2157593002