[Monorail] Lowercase the friendly name to make it look more familiar.
I had gone with an uppercase user name because gmail seems to choose
to display the first uppercase word and I needed the username to
be shown insted of "Monorail".  However, a user said it was odd
looking, and I can also make gmail show the username by keeping
all words lowercase.

R=seanmccullough@chromium.org
BUG=monorail:1465

Review-Url: https://codereview.chromium.org/2173993002