Bootstrap: Enable running even if in a VENV.
Currently, the bootstrap script will quit if it detects that it's
running within a VirtualEnv. Add a flag that allows the script to
continue and back out of the VirtualEnv, --run-within-virtualenv.

Utilize this flag in the "luci-deploy-init" script so the deployment can
be run even if the deployment tool user is currently in a VirtualEnv.

TBR-ing because this will have no impact on current behavior and is part of a set of iterations to get this bootstrap working in a deploytool-initiated environment.

TBR=vadimsh@chromium.org
BUG=None
TEST=local
  - Ran within a VirtualEnv with the flag, environment installed
    successfully.

Review-Url: https://codereview.chromium.org/2178983002