bootstrap: Run "pip" through VENV "python".
Currently, "pip" is run using the "#!/..." shebang notation to run "pip"
through Python. This imposes a limit of 127 characters (by default, on
Linux) for the path to the executable; if this is exceeded, the command
will only recognize the first 127 characters, which means the "pip"
command will fail.

Instead of relying on shebang behavior, explicitly run Pip using the VENV
Python interpreter. This has the same behavior, but avoids the shebang
length limitation.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2177093003