[Monorail] Frontend changes for ranked issue dependencies
A link near the "Blocked on" list in the sidebar opens a table which
shows the rankings of issue dependencies. Users with EditIssue permission
can drag and drop the rows of this table to re-rank the issues. The changes
are automatically sent to the server.

TODO: auto-update other clients who are viewing the table.

BUG=monorail:1458

Review-Url: https://codereview.chromium.org/2183653002