[Monorail] create UserHotlists servlet
This change creates the UserHotlists servlet, which will be used to
serve the "Hotlists" tab on user profiles. It returns 3 lists of
hotlists - owner_of_hotlists, editor_of_hotlists, and
follower_of_hotlists. FOO_of_hotlists is a list of all hotlists
that the viewed user is a FOO of. All hotlists returned are either
public hotlists, or visible to the user viewing the page (i.e. the
viewing user is an owner or editor of that hotlist).

BUG=monorail:50

Change-Id: If752ae0912bfa8c8b65433c9ad6426649226723f
Reviewed-on: https://chromium-review.googlesource.com/370842
Commit-Queue: Lukas Peraza <lukasperaza@google.com>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>