test-results: Update dispatch.yaml (/testfile, /testfile/, /testfile/upload)
When deployed, this will route the paths below to the go module.

 - /testfile
 - /testfile/
 - /testfile/upload

The testfile path appears twice (with and without trailing slash)
to maintain compatibility with a few clients that depend on the path
not 404-ing.

The "*/testfile/" entry can be removed from dispatch.yaml and the path uregistered
from the router in frontend/handlers.go when the client requests are fixed.

BUG=

Review-Url: https://codereview.chromium.org/2256603004