[Monorail] Clarify and simplify issue metadata tooltips.
I used to think it was clever to only show the value in the tooltip if
the value displayed on the page was eliled by the server.  For a
while now, that has been done in CSS, so the tooltip == the name.
This CL eliminates the tooltip and just uses name.

Also, it addresses the issue by always using "=" in front of the docstring.

BUG=monorail:1681

Change-Id: Ibc8647d3a2c50489accfba7a772738b8f33f83d0
Reviewed-on: https://chromium-review.googlesource.com/372798
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>