Install DEPS-homed Go tools in bootstrap.
Currently, there are several tools in "luci" that are useful and/or
required by "go generate", but not installed into PATH by default. This
is because, rather than being vendored via "go/deps.py", they are DEPS'd
in with "gclient".

Add a list of such tools to the "go/bootstrap.py" script, and have it
install those tools whenever either vendoring or Go DEPS versions are
updated. In order to enable this, track the latest DEPS versions for
Go-relevant repositories.

BUG=None
TEST=local

Change-Id: I398d467b0faed0e61664ed6c6aba060252a1960a
Reviewed-on: https://chromium-review.googlesource.com/373918
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>