AD: Refactor reasons
This CL refactors the reason interface to be more general and flexible.

The new JSON generated by the alerts dispatcher is incompatible with
the old JSON in a small way. There is now one reason per alert, rather
than a list of reasons. This means the change will have to be landed
carefully.

Change-Id: Iedb0346a62fb7e2406e5d63b121916f9cffe4329
Reviewed-on: https://chromium-review.googlesource.com/374719
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>