[Monorail] Fix the SQl generated for [-cc:user1,user2].
The pattern to the logic here is that we need to do a LEFT JOIN with
an EQ or TEXT_HAS to attempt to find rows that match the positive form
of the query condition.  Then, we use a WHERE condition with
IS NOT NULL or with IS NULL to see if any such matching value was 
present or absent.

BUG=monorail:1525 and monorail:987

Change-Id: Ib3327986e4b25306d1499e4eaca2384b1bdf516a
Reviewed-on: https://chromium-review.googlesource.com/374659
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>