[Monorail] Allow search by new issue timestamp fields.
This also fixes a defect in my recent CL that uses wildcard values
for the LIKE operator.  I now use string formatting rather than
concatenation to handle the case the value is not a string which
is actually a dont-care, but it should not crash.

BUG=monorail:1070

Change-Id: I736c0724d6c319e39a4dfa4159e7d87ce77a5a6f
Reviewed-on: https://chromium-review.googlesource.com/376181
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>