[Monorail] Display relative dates inside parens again.
In a recent CL I removed the parens around relative dates.  But, that kind of
display is still used on the description and comments on the issue detail page.
So, I added the parens back in the templates, which is a better separation of
logic vs. presentation.  See previous CL:
https://chromium-review.googlesource.com/c/375380/

Change-Id: I31d22052590b92ada2c588df11547f590805d855
Reviewed-on: https://chromium-review.googlesource.com/377744
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>