go: Teach 'deps.py' to bundle all dependencies as a CIPD package.
Currently this is totally optional: it will try to find the bundle in CIPD, but
will fallback to regular Glide if the bundle is not there. It uses SHA1 of
dependencies manifest for identifying versions.

The bundle can be built manually by running 'deps.py bundle'. Eventually this
should run on a waterfall too.

Also fix bug in 'glide install' retries. It was catching the wrong exception.

R=iannucci@chromium.org, dnj@chromium.org
BUG=

Change-Id: Iff6a5e3cd5d91b4d86a04e8829c19afd79e5b68e
Reviewed-on: https://chromium-review.googlesource.com/379477
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>