[Monorail] First steps toward supporting OR queries.
The plan for supporting OR queries is to break the user query into
conjunctions and send each one separately to the backends.  I have
that working in CL 381715, but it is a lot to review.  This CL is a
step toward it that generizes from tracking backend requests by shard_id
to organizing them by shard_key, which will eventually include the
subquery/conjunction.

BUG=monorail:97

Change-Id: Idd07c33130f489be2d5df98f0e6fb9fd051e016e
Reviewed-on: https://chromium-review.googlesource.com/382011
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>