[Monorail] Ignore was_escaped so that we can eventually drop that DB column.
This value was being read and written to the DB, but it was not being
used in the app at all.   It is a very old left-over concept from the
2006 version of code.google.com where we escaped text on input so that
it could be served slightly faster.  That was a bad decision that was
soon reversed, but the scar has lasted until now.

BUG=monorail:1724

Change-Id: I970b79a91583fb407c3cacf430e0364a5ab3e3f5
Reviewed-on: https://chromium-review.googlesource.com/383981
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>