[Monorail] Remove "Delete issue" button that existed for non-JS users.
The "Delete issue" button was put in as a work-around for users who
don't have JS enabled in their browsers.  E.g., in 2005 it was relevant
that basic functionality could work in elinks and emacs's w3 mode.
Eleven years later, this is just cruft, and a user reported seeing
it actually be visible for a second and pressed it.

BUG=monorail:1763

Change-Id: I3e43a2208dd42a5b27656b8d2cb183611cea5ee6
Reviewed-on: https://chromium-review.googlesource.com/384591
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>