[Monorail] Implement OR queries.
As discussed in the issue and CL 382011, OR-queries are supported by
breaking them into multiple backend queries and then merging the
results the same way that we merge shards already.  The concept of
a shard_key is now expanded to be (shard_id, subquery) where each
subquery is one conjunction of the user's query.

The memcache key format is not changed, in fact existing memcache
entries are still valid.

The subquery is not needed or used when computing the set of
non-viewable issue IDs or the staleness timestamps.

BUG=monorail:97

Change-Id: I64f97b70d3d39ec73a182c537df41df3c356a82f
Reviewed-on: https://chromium-review.googlesource.com/383734
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>