[Monorail] Add user_id to the User PB and simplify some functions.
This is a prequel CL to adding a new field to the User PB.  In trying to
add that field, I realized that UserView takes individual arguments, which
would require pervasive changes any time we wanted to add a field.

Also, I added a hook for accepting or ignoring a value found in memcache.
This will allow us to deploy new versions that need new PB fields without
needing to clear memcache at the time of deployment.

BUG=monorail:305

Change-Id: I5c119e213771c11824a2684b0377de03154d87aa
Reviewed-on: https://chromium-review.googlesource.com/383820
Commit-Queue: Jason Robbins <jrobbins+catch-up@google.com>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>