[Findit] Extending versioned_model.py to support versioning multiple entities of the same class.
Previously, versioned_model.py supports creating a singular versioned entity within Findit, and with each update of that singular entity a new version is created and saved. An example of this is wf_config.py, of which there is only 1 entity at any point that is periodically updated.

This change extends versioned_model.py to support creating multiple entities of the same base type by allowing for a string id for the root. For example, MasterFlakeAnalysis is to be versioned, so would have each instance's root key specified as 'master/builder/build_number/step/test' and the integer_id() component responsible for keeping track of which version number is being represented.

BUG=649396

Review-Url: https://codereview.chromium.org/2345093002