Don't save ClientCfgs to datastore if they're invalid
This CL adds lots of error checking in the routine which fetches
new api_client_configs from luci-config. When it runs into an
error, it also increments a new metric representing the attempt
and how/why it failed. This routine is called on a regular basis
by an AppEngine cron task.

BUG=monorail:1688

Change-Id: Id8821f01fe4b19c0e94ac2780c31db16312a04a9
Reviewed-on: https://chromium-review.googlesource.com/390433
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>