[Findit] Factoring out the components, so they can classify themselves
Rather than having ComponentClassifier do a bunch of logic about whether
various components are responsible for various stack frames, we instead
have a new class: Component, which encapsulates the work necessary for
that specific component. That is, each Component can —by itself—
recognize whether it is responsible for a given stack frame (or, in the
future: a given CL, file, etc); and the ComponentClassifier class now
simply stores a list of Components, and calls their methods for them to
decide whether they match or not.

In addition, the ComponentClassifier class no longer depends on
CrashConfig. Instead the constructor is given the values it needs, and
therefore client code is free to get those values from wherever they so
desire (rather than forcing clients to use our internal configuration
tools).

BUG=chromium:605770

Review-Url: https://codereview.chromium.org/2338273006