[Findit] Filter deps which was added/deleted in regression range
If a dep is added/deleted in the regression range, the old_revision/new_revision would be None, in this case, it's very hard to find the regression range of this dep.

There are 3 possiblities:
(1) if the culprit revision got introduced before the dep was added, the dep cannot be responsible for the crash

(2) if the culprit revision got introduced after the dep was added, the regression of this dep should be (introduced_revision, new_revision]

(3) Crashes happened on introduced_revision, this case is complicated, the culprit can be introduced_revision of this dep or any revision before it. (..., introduced_revision]

Since we don't know which revision is the culprit, we cannot know for sure this is the right regression range, besides, this case rarely happens so findit just filters those deps out.

BUG=649143

Review-Url: https://codereview.chromium.org/2363743003