Add git log redirects to crrev
This adds a new redirect condition to cr-rev.appspot.com.
If the query looks like it begins with a git hash, and then
is followed by other random stuff (which we hope is formatted
according to gitrevisions[1]), then we redirect to the repository
containing that starting hash and let gitiles try to parse the
rest of the gitrevision argument.

[1] https://git-scm.com/docs/gitrevisions

BUG=407198

Change-Id: Iedd0b1e2a4dd25c4a875daca7c2a6f6f4d6647d3
Reviewed-on: https://chromium-review.googlesource.com/391014
Commit-Queue: Aaron Gable <agable@chromium.org>
Reviewed-by: Mike Stipicevic <stip@chromium.org>